Fix returning C.CS_INT as int64

+++ cgo/rows.go
@@ -218,7 +218,7 @@ func (rows *rows) ColumnTypePrecisionScale(index int) (int64, int64, bool) {
 func (rows *rows) ColumnTypeLength(index int) (int64, bool) {
 	switch rows.colASEType[index] {
 	case types.BINARY:
-		return rows.dataFmts[index].maxlength, true
+		return int64(rows.dataFmts[index].maxlength), true
 	case types.CHAR:
 		return int64(C.CS_MAX_CHAR), true
 	default:
